Refactor LaravelVuePagination.vue to Bootstrap4Pagination.vue
