+++ src/lib.js
@@ -1,0 +1,7 @@
+import RenderlessPagination from './RenderlessPagination.vue';
+import Bootstrap4Pagination from './Bootstrap4Pagination.vue';
+
+export { 
+    RenderlessPagination, 
+    Bootstrap4Pagination, 
+};
